requirements.txt
beautifulsoup4
google-cloud-secret-manager
pandas
google-cloud-storage